backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-multipart>=0.0.6
